- move components up the tree
