+++ app/src/main/java/com/dsc/myliving/presentation/components/TextInput.kt
@@ -1,10 +1,10 @@
-package com.dsc.myliving.presentation.entry.signin.components
+package com.dsc.myliving.presentation.components
 
 import androidx.compose.material.OutlinedTextField
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import com.dsc.myliving.presentation.entry.signin.state.TextFieldState
-import com.dsc.myliving.presentation.entry.signin.state.rememberTextFieldState
+import com.dsc.myliving.presentation.features.entry.signin.state.TextFieldState
+import com.dsc.myliving.presentation.features.entry.signin.state.rememberTextFieldState
 
 @Composable
 fun TextInput(textFieldState: TextFieldState) {
