+++ app/src/main/java/com/dsc/myliving/presentation/features/entry/EntryActivity.kt
@@ -1,17 +1,13 @@
-package com.dsc.myliving.presentation.entry
+package com.dsc.myliving.presentation.features.entry
 
 import android.content.Intent
 import android.os.Bundle
-import android.os.Handler
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.lifecycle.lifecycleScope
-import com.dsc.myliving.presentation.intro.IntroActivity
-import com.dsc.myliving.presentation.theme.MyLivingTheme
+import com.dsc.myliving.presentation.features.intro.IntroActivity
+import com.dsc.myliving.presentation.features.main.MainActivity
 import com.dsc.myliving.utils.Preferences
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.delay
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -23,11 +19,10 @@ class EntryActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         installSplashScreen()
 
-        // TODO: redirect to auth if token exists
         val token = preferences.getToken()
         val activity = when {
             token != null -> IntroActivity::class.java
-            else -> IntroActivity::class.java
+            else -> MainActivity::class.java
         }
         val intent = Intent(this, activity)
         startActivity(intent)
