+++ app/src/main/java/com/dsc/myliving/presentation/features/entry/signin/screens/SignIn.kt
@@ -1,4 +1,4 @@
-package com.dsc.myliving.presentation.entry.signin.screens
+package com.dsc.myliving.presentation.features.entry.signin.screens
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
